Chatbot/requirements.txt
langchain==0.0.349
openai==1.3.8
validators==0.22.0
python-dotenv
pdfminer
bs4
tiktoken
faiss-cpu
flask
Flask-RESTful